pyproject.toml
[project]
name = "agent-marketplace"
version = "0.1.0"
description = "Router agent that discovers specialist agents via A2A and delegates queries"
authors = [
    {name = "None"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.136.0,<1.0.0)",
    "uvicorn[standard] (>=0.46.0,<1.0.0)",
    "a2a-sdk (>=1.0.0,<2.0.0)",
    "httpx (>=0.28.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "python-multipart (>=0.0.9)",
    "langchain-openai (>=0.3.0,<1.0.0)",
    "langchain (>=1.2.12,<2.0.0)",
    "PyJWT[crypto] (>=2.10.0,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "bcrypt (>=4.0.0)",
    "slowapi>=0.1.9",
    "redis>=5.2.0",
    "agent-sdk @ file:./agent-sdk",
    "azure-keyvault-secrets (>=4.8.0,<5)",
    "azure-identity (>=1.17.0,<2)",
    "prometheus-client (>=0.21.1,<1.0.0)",
    "networkx (>=3.0,<4.0.0)",
    "cachetools (>=7.0.5,<8.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
agent-sdk = { path = "./agent-sdk", develop = true }
agent-common = { path = "./agent-common", develop = true }

[dependency-groups]
dev = [
    "respx (>=0.23.1,<0.24.0)",
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
